==> regresion-lineal/6-pair-regresion-lineal-estandarizacion.ipynb <==
# **Ejercicios de pair programming Módulo 3 Sprint 1**
## **Regresión lineal: Lección 6**
### **Estandarización**

---

En vuestro dataset habréis identificado unas variables predictoras y una variable respuesta. 


# Tratamiento de datos
# ----------------------------------
import pandas as pd
import numpy as np

# Gráficos
# ----------------------------------
import seaborn as sns
import matplotlib.pyplot as plt

# Para realizar la estandarización
# ---------------------------------
from sklearn.preprocessing import StandardScaler

# Para la gestión de los warning
# ----------------------------------
import warnings
warnings.filterwarnings('ignore')

# Para poder ver todas las columnas del dataframe
# ----------------------------------
pd.options.display.max_columns = None

df_socio = pd.read_csv('../datos/datos-regresion-lineal/3-esperanza_vida_normalizado.csv', index_col=0)
df_socio.head(2)


Los objetivos del pair programming de hoy son:
- Sobre las variables predictoras, en este caso deberéis estandarizar estas variables. De nuevo, usad el método que prefiráis.


numericas = df_socio.select_dtypes(include = np.number)
numericas.drop('life_expectancy_at_birth_total_bc', axis = 1, inplace = True)

scaler = StandardScaler()
scaler.fit(numericas)
vp_esc = scaler.transform(numericas)

numericas_estandarizadas = pd.DataFrame(vp_esc, columns = numericas.columns)
numericas_estandarizadas.head(3)

df_socio[numericas.columns] = numericas_estandarizadas

df_socio.head()

- Guardar en un csv el nuevo dataframe que habéis creado para seguir usándolo en los siguientes pair programmings

df_socio.to_csv('../datos/datos-regresion-lineal/4-esperanza_vida_estandarizado.csv')
