==> regresion-lineal/7-pair-regresion-lineal-anova.ipynb <==
# **Ejercicios de pair programming Módulo 3 Sprint 1**
## **Regresión lineal: Lección 7**
### **ANOVA**

---

En el pair programming de hoy usaremos el set de datos que guardastéis en el pair programming de normalización y estandarización.

Hasta ahora habéis estado evaluando las características de vuestro set de datos y habéis hecho una gran exploración, es el momento de hacer vuestro primer ANOVA! En el ejercicio de hoy tendréis que hacer un ANOVA con vuestro datos y hacer una interpretación de los resultados.

📌 NOTA Puede que vuestros datos no se ajusten o no cumplan todas las asunciones, no pasa nada, haced el ANOVA e interpretad los resultados. En próximas lecciones aprenderemos que podemos hacer cuando nos encontramos en esta situación.

# Tratamiento de datos
# ------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Estadísticos
# ------------------------------
import statsmodels.api as sm
from statsmodels.formula.api import ols
from statsmodels.multivariate.manova import MANOVA
from sklearn.preprocessing import StandardScaler

plt.rcParams["figure.figsize"] = (10,8)

df_socio = pd.read_csv('../datos/datos-regresion-lineal/4-esperanza_vida_estandarizado.csv', index_col=0)
df_socio.head(2)

anova = ols("""life_expectancy_at_birth_total_bc ~ region + population_in_millions + population_density +
       urban_population_y + gdp_per_capita + health_total_expenditure +
       health_care_index + quantity_of_physicians +
       quality_of_life_index + purchasing_power_index + safety_index +
       pop_using_improved_drinking_water + pollution_index""", data = df_socio).fit()

sm.stats.anova_lm(anova)

anova.summary()

> Podemos observar que el R cuadrado se sitúa en un 0.891, lo que indica que el 89.1% de nuestra variable respuesta es causado por nuestras variables predictoras.
> 
> Todos los valores de nuestra columna PR(>F) están por debajo de 0.05, por lo que podemos asumir que todas las variables predictoras afectan a la variable respuesta.
