==> regresion-lineal/8-pair-regresion-lineal-encoding.ipynb <==
# **Ejercicios de pair programming Módulo 3 Sprint 1**
## **Regresión lineal: Lección 8**
### **Encoding**

---

# Tratamiento de datos
# ------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Para la codificación de las variables numéricas
# ------------------------------
from sklearn.preprocessing import OrdinalEncoder  

# Para evitar que salgan los warnings en jupyter
# ------------------------------
import warnings
warnings.filterwarnings('ignore')

pd.options.display.max_columns = None

En el pair programming de hoy usaremos el set de datos que guardastéis en el pair programming de normalización y estandarización.
Vuestro set de datos debería tener al menos una variable categórica, el objetivo del pair programming de hoy:


df_socio = pd.read_csv('../datos/datos-regresion-lineal/4-esperanza_vida_estandarizado.csv', index_col=0)
df_socio.head(2)


 - Hacer una códificación de la/las variables categóricas que tengáis en vuestro set de datos. Recordad que lo primero que deberéis hacer es decidir su vuestras variables tienen o no orden, para que en función de esto uséis una aproximación u otra.


# Establecemos la columna 'country' como nuestro índice, ya que no queremos perder esta información pero no se trata de una variable predictora realmente.
# ------------------------------

df_socio = df_socio.set_index("country")

df_socio.dtypes

> Nuestras dos variables categóricas ('region' y 'quantity_of_physicians') tienen orden, de modo que realizamos el encoding con 

# Hacemos primero el encoding de la variable referente a la cantidad de médicos.
# ------------------------------

diccionario_medicos = {'Muchos': 3, 'Medios': 2.5, 'Muy pocos': 0, 'Pocos': 1}

df_socio["quantity_of_physicians"] = df_socio["quantity_of_physicians"].map(diccionario_medicos)

# Por último hacemos el encoding de la variable 'region'.
# Decidimos realizar dos encodings diferentes y por tanto guardamos dos dataframes distintos.
# ------------------------------

diccionario_eli = {'SouthAmerica': 2, 'Oceania': 3, 'WesternEurope': 3, 'EasternEurope': 2,
       'SouthernEurope': 3, 'NorthernAmerica': 3, 'EasternAsia': 3, 'WesternAsia': 2,
       'NorthernEurope': 3, 'NorthernAfrica': 1, 'SouthernAsia': 1,
       'South-easternAsia': 2, 'CentralAmerica': 2, 'SouthernAfrica':0}

diccionario_lidya = {'SouthAmerica': 4, 'Oceania': 5, 'WesternEurope': 5, 'EasternEurope': 3,
       'SouthernEurope': 5, 'NorthernAmerica': 5, 'EasternAsia': 5, 'WesternAsia': 4,
       'NorthernEurope': 5, 'NorthernAfrica': 2, 'SouthernAsia': 3,
       'South-easternAsia': 3, 'CentralAmerica': 3, 'SouthernAfrica': 0}

df_socio_eli = df_socio.copy()
df_socio_lidya = df_socio.copy()

df_socio_eli["region"] = df_socio_eli["region"].map(diccionario_eli)
df_socio_lidya["region"] = df_socio_lidya["region"].map(diccionario_lidya)

df_socio_eli.head(2)

df_socio_lidya.head(2)


 - Guardad el dataframe, donde deberíais tener las variables estadandarizas, normalizadas y codificadas en un csv para usarlo en el próximo pairprogramming

df_socio_eli.to_csv("../datos/datos-regresion-lineal/5.1-esperanza_vida_encoding_eli.csv")
df_socio_lidya.to_csv("../datos/datos-regresion-lineal/5.2-esperanza_vida_encoding_lidya.csv")
