==> regresion-lineal/10-pair-regresion-lineal-metricas.ipynb <==
# **Ejercicios de pair programming Módulo 3 Sprint 1**
## **Regresión lineal: Lección 10**
### **Regresión lineal: Métricas**

---

# Tratamiento de datos
# -----------------------------------------------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

#  Modelado y evaluación
# -----------------------------------------------------------------------------
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error

# Para navegar entre directorios
# -----------------------------
import os

plt.rcParams["figure.figsize"] = (10,8)

# Para evitar que salgan los warnings en jupyter
# -----------------------------------------------------------------------
import warnings
warnings.filterwarnings('ignore')

pd.options.display.max_columns = None

En el pair programming anterior creastéis vuestro primer modelo de machine learning usando la regresion Lineal. Es el momento, que con vuestros datos evaluéis si es bueno haciendo predicciones. Los objetivo de este pairprogramming son:


os.getcwd()

os.chdir('/mnt/c/Users/lidya/Documents/Nuestros documentos/Documentos de Lidya/Estudios/Adalab/modulos/modulo3/pair_programming/promo-d-da-modulo3-sprint1-elisa-lidya/regresion-lineal')

df_socio = pd.read_csv('../datos/esperanza_vida_transformado.csv', index_col=0)
df_socio.head(2)

 - Calculéis las métricas para vuestro modelo.


# Volvemos a hacer la regresión del pair anterior.

y = df_socio['life_expectancy_at_birth_total_bc']
x = df_socio.drop('life_expectancy_at_birth_total_bc', axis=1)

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.3, random_state = 42)

lr = LinearRegression(n_jobs = -1)

lr.fit(x_train, y_train)

y_predict_train = lr.predict(x_train)
y_predict_test = lr.predict(x_test)

df_train = pd.DataFrame({'real' : y_train, 'predicted' : y_predict_train, 'set' : ['train']* len(y_train)})
df_test = pd.DataFrame({'real' : y_test, 'predicted' : y_predict_test, 'set' : ['test']* len(y_test)})

resultado = pd.concat([df_train, df_test], axis=0)

resultado.head()

resultados_metricas = {'MAE': [mean_absolute_error(y_test, y_predict_test), mean_absolute_error(y_train, y_predict_train)],
                'MSE': [mean_squared_error(y_test, y_predict_test), mean_squared_error(y_train, y_predict_train)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_predict_test)), np.sqrt(mean_squared_error(y_train, y_predict_train))],
                'R2':  [r2_score(y_test, y_predict_test), r2_score(y_train, y_predict_train)],
                 "set": ["test", "train"], 
                 "modelo": ["LinearRegression", "LinearRegression"]}

df_resultados = pd.DataFrame(resultados_metricas)

df_resultados

 - Discutid los resultados de las métricas y extraed conclusiones.


~~~

Vamos a analizar las métricas por tipo:

- R2: Podemos observar que el R2 del conjunto de entrenamiento más que duplica el R2 del conjunto del test, lo que puede sugerir que el modelo sufre un problema de *overfitting*.

- MAE, MSE, RMSE: Estas tres métricas tienen valores bastante bajos, casi 0, y son muy similares en los tres casos para los dos conjuntos de datos. Aún así podemos comprobar, por ejemplo, que el RMSE del test sigue siendo más del doble que el del train. 

- Podemos ver en general que las métricas se relacionan de forma más o menos estable en sus diferencias. 

~~~


 - Guardad los resultados de las métricas en un csv para usarlo más adelante.

df_resultados.to_csv("../datos/resultados_rl.csv")
