==> regresion-lineal/11.2-pair-regresion-lineal-decision-tree-lidya.ipynb <==
# **Ejercicios de pair programming Módulo 3 Sprint 1**
## **Regresión lineal: Lección 11**
### **Decision Tree**

---

# Tratamiento de datos
# ------------------------------------------------------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Modelado y evaluación
# ------------------------------------------------------------------------------
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV

# Configuración warnings
# ------------------------------------------------------------------------------
import warnings
warnings.filterwarnings('once')

pd.options.display.max_columns = None

En el pair programming de hoy debéis usar el csv que guardastéis cuando hicistéis el pairprgramming de codificicación (este csv debería tener las variables estadandarizas).

df_socio = pd.read_csv('../datos/datos-regresion-lineal/5.2-esperanza_vida_encoding_lidya.csv', index_col=0)
df_socio.head(2)


En pairprogramming anteriores ajustastéis vuestro datos a una regresión lineal. El objetivo de hoy es:
 - Ajustar el modelo a un Decision Tree


X = df_socio.drop('life_expectancy_at_birth_total_bc', axis=1)
y = df_socio['life_expectancy_at_birth_total_bc']

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, random_state = 63)

arbol = DecisionTreeRegressor(random_state = 137)

arbol.fit(x_train, y_train)

max_features = np.sqrt(len(x_train.columns))
print(f'La máxima cantidad de features que podemos incluir en el modelo es {max_features}.')
print(f'La máxima profundidad que puede tener nuestro modelo es {arbol.tree_.max_depth}')

y_pred_test = arbol.predict(x_test)
y_pred_train = arbol.predict(x_train)

def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    
    
    resultados = {'MAE': [mean_absolute_error(y_test, y_test_pred), mean_absolute_error(y_train, y_train_pred)],
                'MSE': [mean_squared_error(y_test, y_test_pred), mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_test_pred)), np.sqrt(mean_squared_error(y_train, y_train_pred))],
                'R2':  [r2_score(y_test, y_test_pred), r2_score(y_train, y_train_pred)],
                "set": ["test", "train"]}
    
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

resultados1 = metricas(y_test, y_train, y_pred_test, y_pred_train, 'Arbol de decisión Lidya 1')

resultados1

>Vemos que tenemos un claro caso de overfitting, así que utilizamos GridSearchCV para ajustar nuestro hiperparámetros.

print(f'La máxima cantidad de features que podemos incluir en el modelo es {max_features}.')
print(f'La máxima profundidad que puede tener nuestro modelo es {arbol.tree_.max_depth}')

param = {"max_depth": [2,3,4,6],  
        "max_features": [3,4,5,6],        
        "min_samples_split": [10,50,100],
        "min_samples_leaf": [10,50,100]}

gs = GridSearchCV(
            estimator=DecisionTreeRegressor(), 
            param_grid= param, 
            cv=10, 
            verbose=-1, 
            return_train_score = True, 
            scoring="neg_mean_squared_error")

gs.fit(x_train, y_train)

modelo1 = gs.best_estimator_
modelo1

param2 = {"max_depth": [2,3,4,6],  
        "max_features": [3,4,5,6],    
        "min_samples_split": [2,4,5,10,15],
        "min_samples_leaf": [2,4,5,10,15]}

gs2 = GridSearchCV(
            estimator=DecisionTreeRegressor(), 
            param_grid= param2, 
            cv=10, 
            verbose=-1, 
            return_train_score = True, 
            scoring="neg_mean_squared_error")

gs2.fit(x_train, y_train)

modelo2 = gs2.best_estimator_
modelo2

#volvemos a hacer el método con los hiperparámetros que nos ha devuelto el GridSearchCV

x_train2, x_test2, y_train2, y_test2 = train_test_split(X, y, test_size = 0.3, random_state = 63)

arbol2 = DecisionTreeRegressor(max_depth=3, max_features=3, min_samples_leaf=10,
                      min_samples_split=10)

arbol2.fit(x_train2, y_train2)

y_pred_test2 = arbol2.predict(x_test2)
y_pred_train2 = arbol2.predict(x_train2)

resultados2 = metricas(y_test2, y_train2, y_pred_test2, y_pred_train2, 'Arbol de decisión Lidya 2')
resultados2

importancia_predictores = pd.DataFrame({'predictor': x_train2.columns,'importancia': modelo2.feature_importances_})
importancia_predictores.sort_values(by=["importancia"], ascending=False, inplace = True)
print("Importancia de los predictores en el modelo")
print("-------------------------------------------")
importancia_predictores

 - Extraer las métricas


resultados = pd.read_csv("../datos/datos-regresion-lineal/6.2-esperanza_vida_metricas_lidya.csv", index_col =0)

resultados = pd.concat([resultados, resultados1, resultados2], axis = 0)

resultados

 - Debatid entre vosotras que modelo es mejor y por qué (basándose en las métricas)

>Ninguno de los modelos que hemos intentando hacer hasta ahora nos ha dado un resultado satisfactorio.
> El mejor de ellos es el segundo árbol de decisión, y aún así éste sufre un enorme overfitting.

resultados.to_csv("../datos/datos-regresion-lineal/6.2-esperanza_vida_metricas_lidya.csv")
